move Frame to transform.ts
