+++ src/phys_esc.ts
@@ -28,6 +28,7 @@ import {
 } from "./phys_broadphase.js";
 import { moveObjects } from "./phys_motion.js";
 import {
+  Frame,
   ParentTransformDef,
   Position,
   PositionDef,
@@ -53,33 +54,6 @@ import { assert } from "./test.js";
 //      otherwise we have a "recomputeTransform" or "recomputeYYY" to rebuild parts from the others
 //      ideally there would be some flags to help know what is out of sync
 //    is it possible Frames could be read-only and all mutation could be pushed back to the "truth" fields?
-
-export interface Frame {
-  transform: mat4;
-  position: vec3;
-  rotation: quat;
-  scale: vec3;
-}
-
-export function updateFrameFromTransform(
-  f: Partial<Frame>
-): asserts f is Frame {
-  f.transform = f.transform ?? mat4.create();
-  f.position = mat4.getTranslation(f.position ?? vec3.create(), f.transform);
-  f.rotation = mat4.getRotation(f.rotation ?? quat.create(), f.transform);
-  f.scale = mat4.getScaling(f.scale ?? vec3.create(), f.transform);
-}
-
-export function updateFrameFromPosRotScale(
-  f: Partial<Frame>
-): asserts f is Partial<Frame> & { transform: mat4 } {
-  f.transform = mat4.fromRotationTranslationScale(
-    f.transform ?? mat4.create(),
-    f.rotation ?? quat.IDENTITY,
-    f.position ?? vec3.ZEROS,
-    f.scale ?? vec3.ONES
-  );
-}
 
 export const PhysicsResultsDef = EM.defineComponent("physicsResults", () => {
   return {
