+++ src/phys_motion.ts
@@ -2,7 +2,8 @@ import { Collider } from "./collider.js";
 import { quat, vec3 } from "./gl-matrix.js";
 import { clamp } from "./math.js";
 import { IdPair, ContactData } from "./phys.js";
-import { Frame, PhysicsState } from "./phys_esc.js";
+import { PhysicsState } from "./phys_esc.js";
+import { Frame } from "./transform.js";
 
 let delta = vec3.create();
 let normalizedVelocity = vec3.create();
